Refactor googleAPI validity checks & add reschedule
Now checks for request validity shouldn't take so long and avoid
some edge cases. Room being free is now checked before making the
request, which is not 100% reliable because someone else can make a
booking in the time between the check and the actual request.

This is still better than checking afterwards, because accepting can take a lot of time
and many more requests are required to revert all changes with their
own chances of failing.